Major Actor __init__() Refactor
Most if not all actors will be affected by this change.

The super().__init__() used in the Actor class should now go at the very top of the function. A new Cunstruct() method has been created. This is where you will create components and use attributes that are created after __init__ but before it is spawned in the world (Mostly things that involve rect and position)